code/src/requirements.txt
pdfplumber
google_generativeai
bs4
google-cloud-storage
google-cloud-vision
